--- src/main/java/hongik/heavyYoun/HeavyYoungApplication.java
@@ -1,13 +1,0 @@
-package hongik.heavyYoun;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HeavyYoungApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HeavyYoungApplication.class, args);
-    }
-
-}
