--- src/main/java/hongik/heavyYoun/domain/item/entity/ItemCategory.java
@@ -1,32 +1,0 @@
-// src/main/java/com/yourapp/domain/item/entity/ItemCategory.java
-package hongik.heavyYoun.domain.item.entity;
-
-import hongik.heavyYoun.domain.item.enums.ItemCategoryName;
-import hongik.heavyYoun.global.baseEntity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Entity
-@Table(name = "item_category")
-public class ItemCategory extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "item_category_id")
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "category_name", nullable = false, length = 20)
-    private ItemCategoryName categoryName;
-
-    @Column(name = "total_quantity", nullable = false)
-    @Builder.Default
-    private Integer totalQuantity = 0;
-
-    @Column(name = "current_quantity", nullable = false)
-    @Builder.Default
-    private Integer currentQuantity = 0;
-}
