--- src/main/java/hongik/heavyYoun/domain/item/enums/ItemCategoryName.java
@@ -1,7 +1,0 @@
-package hongik.heavyYoun.domain.item.enums;
-
-public enum ItemCategoryName {
-    UMBRELLA, // 우산
-    BATTERY, // 보조 배터리
-    LAPTOP_CHARGER, // 노트북 충전기
-}
