--- src/main/java/hongik/heavyYoun/domain/locker/enums/LockerApplicationPeriodType.java
@@ -1,6 +1,0 @@
-package hongik.heavyYoun.domain.locker.enums;
-
-public enum LockerApplicationPeriodType {
-    MAIN, // 정규 신청
-    ADDITIONAL, // 추가 신청
-}
