--- src/main/java/hongik/heavyYoun/domain/locker/enums/LockerStatus.java
@@ -1,7 +1,0 @@
-package hongik.heavyYoun.domain.locker.enums;
-
-public enum LockerStatus {
-    BROKEN, // 고장
-    IN_USE, // 대여중
-    AVAILABLE, // 사용 가능
-}
