--- src/main/java/hongik/heavyYoun/domain/member/enums/MemberRole.java
@@ -1,7 +1,0 @@
-package hongik.heavyYoun.domain.member.enums;
-
-public enum MemberRole {
-    USER, // 일반 학생
-    ADMIN, // 학생회 학생
-    OWNER, // 학생회 마스터 계정
-}
