--- src/main/java/hongik/heavyYoun/domain/member/enums/MemberStatus.java
@@ -1,7 +1,0 @@
-package hongik.heavyYoun.domain.member.enums;
-
-public enum MemberStatus {
-    PENDING, // 회원 가입 중
-    ACTIVE, // 활성화
-    DELETED, // 회원 탈퇴
-}
