+++ src/main/java/hongik/heavyYoung/HeavyYoungApplication.java
@@ -1,0 +1,13 @@
+package hongik.heavyYoung;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HeavyYoungApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HeavyYoungApplication.class, args);
+    }
+
+}
