+++ src/main/java/hongik/heavyYoung/domain/event/entity/EventImage.java
@@ -1,0 +1,26 @@
+package hongik.heavyYoung.domain.event.entity;
+
+import hongik.heavyYoung.global.baseEntity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity
+@Table(name = "event_image")
+public class EventImage extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "event_image_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "event_id", nullable = false)
+    private Event event;
+
+    @Lob
+    @Column(name = "event_image_url", nullable = false)
+    private String eventImageUrl;
+}
