+++ src/main/java/hongik/heavyYoung/domain/item/entity/Item.java
@@ -1,0 +1,26 @@
+package hongik.heavyYoung.domain.item.entity;
+
+import hongik.heavyYoung.global.baseEntity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity
+@Table(name = "item")
+public class Item extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "item_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "item_category_id", nullable = false)
+    private ItemCategory itemCategory;
+
+    @Column(name = "rental_status", nullable = false)
+    @Builder.Default
+    private Boolean rentalStatus = false;
+}
