+++ src/main/java/hongik/heavyYoung/domain/item/enums/ItemCategoryName.java
@@ -1,0 +1,7 @@
+package hongik.heavyYoung.domain.item.enums;
+
+public enum ItemCategoryName {
+    UMBRELLA, // 우산
+    BATTERY, // 보조 배터리
+    LAPTOP_CHARGER, // 노트북 충전기
+}
