+++ src/main/java/hongik/heavyYoung/domain/locker/entity/LockerApplication.java
@@ -1,0 +1,27 @@
+package hongik.heavyYoung.domain.locker.entity;
+
+import hongik.heavyYoung.domain.member.entity.Member;
+import hongik.heavyYoung.global.baseEntity.BaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity
+@Table(name = "locker_application")
+public class LockerApplication extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "locker_application_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "locker_application_period", nullable = false)
+    private LockerApplicationPeriod lockerApplicationPeriod;
+}
