+++ src/main/java/hongik/heavyYoung/domain/locker/enums/LockerApplicationPeriodType.java
@@ -1,0 +1,6 @@
+package hongik.heavyYoung.domain.locker.enums;
+
+public enum LockerApplicationPeriodType {
+    MAIN, // 정규 신청
+    ADDITIONAL, // 추가 신청
+}
