+++ src/main/java/hongik/heavyYoung/domain/locker/enums/LockerStatus.java
@@ -1,0 +1,7 @@
+package hongik.heavyYoung.domain.locker.enums;
+
+public enum LockerStatus {
+    BROKEN, // 고장
+    IN_USE, // 대여중
+    AVAILABLE, // 사용 가능
+}
