+++ src/main/java/hongik/heavyYoung/domain/member/enums/MemberRole.java
@@ -1,0 +1,7 @@
+package hongik.heavyYoung.domain.member.enums;
+
+public enum MemberRole {
+    USER, // 일반 학생
+    ADMIN, // 학생회 학생
+    OWNER, // 학생회 마스터 계정
+}
