+++ src/main/java/hongik/heavyYoung/domain/member/enums/MemberStatus.java
@@ -1,0 +1,7 @@
+package hongik.heavyYoung.domain.member.enums;
+
+public enum MemberStatus {
+    PENDING, // 회원 가입 중
+    ACTIVE, // 활성화
+    DELETED, // 회원 탈퇴
+}
