+++ src/main/java/hongik/heavyYoung/domain/member/enums/StudentFeeStatus.java
@@ -1,0 +1,7 @@
+package hongik.heavyYoung.domain.member.enums;
+
+public enum StudentFeeStatus {
+    PAID, // 납부 완료
+    NOT_PAID, // 미납
+    YET, // 엑셀 동기화 필요
+}
