--- src/test/java/hongik/heavyYoun/HeavyYoungApplicationTests.java
@@ -1,13 +1,0 @@
-package hongik.heavyYoun;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HeavyYoungApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
