+++ src/test/java/hongik/heavyYoung/HeavyYoungApplicationTests.java
@@ -1,0 +1,13 @@
+package hongik.heavyYoung;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HeavyYoungApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
